app/pyproject.toml
[tool.poetry]
name = "wic-mock-api"
version = "0.1.0"
description = "A Mock API wrapper around the SPIRIT MIS for the Special Supplemental Nutrition Program for Women, Infants, and Children (WIC)"
packages = [{ include = "api" }]
authors = ["Nava Engineering <engineering@navapbc.com>"]

[tool.poetry.dependencies]
python = "^3.10"
connexion = {extras = ["swagger-ui"], version = "^2.13.1"}
SQLAlchemy = "^1.4.40"
Faker = "^14.1.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
black = "^22.6.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.8.23"
flake8-alfred = "^1.1.1"
isort = "^5.10.1"
mypy = "^0.971"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
db-migrate-up = "api.db.migrations.run:up"
db-migrate-down = "api.db.migrations.run:down"
db-migrate-down-all = "api.db.migrations.run:downall"

[tool.black]
line-length = 100

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100

[tool.mypy]
color_output = true
error_summary = true
ignore_missing_imports = true
namespace_packages = true
pretty = true
show_column_numbers = true
show_error_context = true
warn_unused_configs = true

check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_equality = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true
